requirements.txt
Flask==1.1.1
Flask-SocketIO==4.2.1
protobuf==3.11.1
pycryptodome==3.9.4
pytest-cov==2.8.1
requests==2.22.0
tqdm==4.40.1
